# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SchoolLevel": {
				"description": "Enum representing different school levels.\n0 = Elementary School, 1 = Middle School, 2 = High School",
				"enum": [
					"0",
					"1",
					"2"
				],
				"type": "string"
			},
...
			"RegistrationType": {
				"description": "Enum representing different types of registration.\n0 = Zoning, 1 = Achievement, 2 = Affirmation, 3 = Transfer, 4 = Custom",
				"enum": [
					"0",
					"1",
					"2",
					"3",
					"4"
				],
				"type": "string"
			},
...
			"RegistrationPathStatus": {
				"description": "0 = Active, 1 = Non-active",
				"enum": [
					"1",
					"0"
				],
				"type": "string"
			},
			"RegistrationPath": {
				"properties": {
					"is_have_payment": {
						"type": "boolean",
						"description": "Boolean value indicating whether the registration path has payment or not."
					},
					"status": {
						"$ref": "#/components/schemas/RegistrationPathStatus",
						"description": "Registration path status"
					},
					"end_date": {
						"type": "string",
						"description": "ISO 8601 date-time string"
					},
					"start_date": {
						"type": "string",
						"description": "ISO 8601 date-time string"
					},
					"description": {
						"type": "string"
					},
					"registration_type": {
						"$ref": "#/components/schemas/RegistrationType"
					},
					"level": {
						"$ref": "#/components/schemas/SchoolLevel"
					},
					"name": {
						"type": "string",
						"description": "Name of the registration path"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"is_have_payment",
					"status",
					"end_date",
					"start_date",
					"description",
					"registration_type",
					"level",
					"name",
					"id"
				],
				"type": "object"
			},
			"PaginatedResponseEnvelope_RegistrationPath_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "HTTP Response code"
					},
					"error": {
						"type": "boolean",
						"description": "Error flag"
					},
					"message": {
						"type": "string",
						"description": "Message"
					},
					"count": {
						"type": "number",
						"format": "double",
						"description": "Total number of items"
					},
					"next": {
						"type": "string",
						"description": "Url of the next page"
					},
					"previous": {
						"type": "string",
						"description": "Url of the previous page"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/RegistrationPath"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"error",
					"message",
					"count",
					"next",
					"previous",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseEnvelope_RegistrationPath_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "HTTP Response code"
					},
					"error": {
						"type": "boolean",
						"description": "Error flag"
					},
					"message": {
						"type": "string",
						"description": "Message"
					},
					"data": {
						"$ref": "#/components/schemas/RegistrationPath"
					}
				},
				"required": [
					"code",
					"error",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegistrationPathFieldType": {
				"description": "0 = Text, 1 = Number, 2 = Choice, 3 = Multiple Choice, 4 = Date, 5 = Document",
				"enum": [
					"0",
					"1",
					"2",
					"3",
					"4",
					"5"
				],
				"type": "string"
			},
			"RegistrationPathConfig": {
				"properties": {
					"is_editable": {
						"type": "boolean"
					},
					"is_deleteable": {
						"type": "boolean"
					},
					"fields": {
						"items": {
							"properties": {
								"point": {
									"type": "number",
									"format": "double",
									"nullable": true,
									"description": "Field\\`s point, nullable, null if the field`s type its not a Document."
								},
								"is_editable": {
									"type": "boolean"
								},
								"ordering": {
									"type": "number",
									"format": "double",
									"description": "Field ordering"
								},
								"options": {
									"items": {
										"properties": {
											"label": {
												"type": "string",
												"description": "Option label"
											},
											"id": {
												"type": "number",
												"format": "double",
												"description": "Option ID"
											}
										},
										"required": [
											"label"
										],
										"type": "object"
									},
									"type": "array",
									"description": "List of options for the field. Only applicable for field type 2 and 3."
								},
								"field_type": {
									"$ref": "#/components/schemas/RegistrationPathFieldType",
									"description": "Field type\n0 = Text, 1 = Number, 2 = Choice, 3 = Multiple Choice, 4 = Date, 5 = Document"
								},
								"is_required": {
									"type": "boolean",
									"description": "Boolean value indicating whether the field is required or not."
								},
								"title": {
									"type": "string",
									"description": "Field title"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"point",
								"is_editable",
								"ordering",
								"options",
								"field_type",
								"is_required",
								"title"
							],
							"type": "object"
						},
						"type": "array",
						"description": "List of fields in the group"
					},
					"name": {
						"type": "string",
						"description": "Fields group name"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"is_editable",
					"is_deleteable",
					"fields",
					"name"
				],
				"type": "object"
			},
			"RegistrationPathBillingScheduleType": {
				"description": "0 = Pay at registration, 1 = Pay after acceptance",
				"enum": [
					"0",
					"1"
				],
				"type": "string"
			},
			"RegistrationPathPaymentMethod": {
				"enum": [
					"BANK_TRANSFER",
					"QRIS",
					"EWALLET"
				],
				"type": "string"
			},
			"RegistrationPathPaymentConfig": {
				"properties": {
					"payment_methods": {
						"items": {
							"$ref": "#/components/schemas/RegistrationPathPaymentMethod"
						},
						"type": "array",
						"description": "List of payment methods available for the registration path."
					},
					"items": {
						"items": {
							"properties": {
								"billing_schedule_type": {
									"$ref": "#/components/schemas/RegistrationPathBillingScheduleType",
									"description": "When the payment should be made.\n0 = Pay at registration, 1 = Pay after acceptance"
								},
								"price": {
									"type": "number",
									"format": "double"
								},
								"name": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"billing_schedule_type",
								"price",
								"name"
							],
							"type": "object"
						},
						"type": "array",
						"description": "List of items need to be paid for the registration path."
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payment_methods",
					"items"
				],
				"type": "object",
				"description": "Payment configuration for the registration path (Step 4)"
			},
			"RegistrationPathPayload": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"general": {
						"properties": {
							"end_date": {
								"type": "string",
								"description": "ISO 8601 date-time string"
							},
							"start_date": {
								"type": "string",
								"description": "ISO 8601 date-time string"
							},
							"description": {
								"type": "string"
							},
							"registration_type": {
								"$ref": "#/components/schemas/RegistrationType"
							},
							"level": {
								"$ref": "#/components/schemas/SchoolLevel"
							},
							"name": {
								"type": "string",
								"description": "Name of the registration path"
							}
						},
						"required": [
							"end_date",
							"start_date",
							"description",
							"registration_type",
							"level",
							"name"
						],
						"type": "object",
						"description": "Registration path general information (Step 1)"
					},
					"configuration": {
						"items": {
							"$ref": "#/components/schemas/RegistrationPathConfig"
						},
						"type": "array",
						"description": "Form configuration for the registration path (Step 2)"
					},
					"schools": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "List of school IDs that are associated with the registration path (Step 3)"
					},
					"payment_config": {
						"$ref": "#/components/schemas/RegistrationPathPaymentConfig"
					},
					"zone_id": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Registration path zone id, null if the registration path is not zone-based."
					}
				},
				"required": [
					"general",
					"configuration",
					"schools",
					"zone_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegistrationPathDetail": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"general": {
						"properties": {
							"end_date": {
								"type": "string",
								"description": "ISO 8601 date-time string"
							},
							"start_date": {
								"type": "string",
								"description": "ISO 8601 date-time string"
							},
							"description": {
								"type": "string"
							},
							"registration_type": {
								"$ref": "#/components/schemas/RegistrationType"
							},
							"level": {
								"$ref": "#/components/schemas/SchoolLevel"
							},
							"name": {
								"type": "string",
								"description": "Name of the registration path"
							}
						},
						"required": [
							"end_date",
							"start_date",
							"description",
							"registration_type",
							"level",
							"name"
						],
						"type": "object",
						"description": "Registration path general information (Step 1)"
					},
					"configuration": {
						"items": {
							"$ref": "#/components/schemas/RegistrationPathConfig"
						},
						"type": "array",
						"description": "Form configuration for the registration path (Step 2)"
					},
					"schools": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "List of school IDs that are associated with the registration path (Step 3)"
					},
					"payment_config": {
						"$ref": "#/components/schemas/RegistrationPathPaymentConfig"
					},
					"zone_id": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Registration path zone id, null if the registration path is not zone-based."
					},
					"has_applicant": {
						"type": "boolean",
						"description": "Boolean value indicating whether the registration path has applicant or not."
					}
				},
				"required": [
					"general",
					"configuration",
					"schools",
					"zone_id",
					"has_applicant"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseEnvelope_RegistrationPathDetail_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "HTTP Response code"
					},
					"error": {
						"type": "boolean",
						"description": "Error flag"
					},
					"message": {
						"type": "string",
						"description": "Message"
					},
					"data": {
						"$ref": "#/components/schemas/RegistrationPathDetail"
					}
				},
				"required": [
					"code",
					"error",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseEnvelope_null_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "HTTP Response code"
					},
					"error": {
						"type": "boolean",
						"description": "Error flag"
					},
					"message": {
						"type": "string",
						"description": "Message"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"code",
					"error",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegistrationPathPaymentStatus": {
				"enum": [
					"CREATED",
					"WAITING_PAYMENT",
					"PAYMENT_IN_PROGRESS",
					"DISBURSE_IN_PROGRESS",
					"COMPLETE",
					"INCOMPLETE",
					"EXPIRED",
					"PAYMENT_FAILED"
				],
				"type": "string"
			},
			"RegistrationPathPayment": {
				"properties": {
					"transaction_code": {
						"type": "string"
					},
					"payment_status": {
						"$ref": "#/components/schemas/RegistrationPathPaymentStatus"
					},
					"payment_method": {
						"type": "string"
					},
					"payment_date": {
						"type": "string"
					},
					"payment_amount": {
						"type": "number",
						"format": "double"
					},
					"fee_names": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"student_name": {
						"type": "string"
					},
					"billing_schedule_type": {
						"$ref": "#/components/schemas/RegistrationPathBillingScheduleType"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"transaction_code",
					"payment_status",
					"payment_method",
					"payment_date",
					"payment_amount",
					"fee_names",
					"student_name",
					"billing_schedule_type",
					"id"
				],
				"type": "object"
			},
			"PaginatedResponseEnvelope_RegistrationPathPayment_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "HTTP Response code"
					},
					"error": {
						"type": "boolean",
						"description": "Error flag"
					},
					"message": {
						"type": "string",
						"description": "Message"
					},
					"count": {
						"type": "number",
						"format": "double",
						"description": "Total number of items"
					},
					"next": {
						"type": "string",
						"description": "Url of the next page"
					},
					"previous": {
						"type": "string",
						"description": "Url of the previous page"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/RegistrationPathPayment"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"error",
					"message",
					"count",
					"next",
					"previous",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "PPDB API",
		"version": "1.0.0",
		"description": "PPDB API",
		"license": {
			"name": "Private"
		},
		"contact": {}
	},
	"paths": {
...
		"/v2/ppdb/registration-paths": {
			"get": {
				"operationId": "GetRegistrationPaths",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseEnvelope_RegistrationPath_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": 200,
											"error": false,
											"message": "Success",
											"count": 100,
											"next": "",
											"previous": "",
											"data": [
												{
													"id": 1,
													"name": "Path 1",
													"level": "0",
													"registration_type": "1",
													"description": "This is the first path",
													"start_date": "2021-01-01T00:00:00Z",
													"end_date": "2021-01-01T00:00:00Z",
													"status": "1",
													"is_have_payment": false
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get All Registration Paths paginated",
				"summary": "Get all registration paths",
				"tags": [
					"Registration Path"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page_size",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Array of SchoolLevel",
						"in": "query",
						"name": "levels",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/SchoolLevel"
							}
						}
					},
					{
						"in": "query",
						"name": "registration_types",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/RegistrationType"
							}
						}
					}
				]
			},
			"post": {
				"operationId": "CreateRegistrationPath",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseEnvelope_RegistrationPath_"
								}
							}
						}
					}
				},
				"description": "Create a new Registration Path",
				"summary": "Create a new registration path",
				"tags": [
					"Registration Path"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegistrationPathPayload"
							}
						}
					}
				}
			}
		},
		"/v2/ppdb/registration-paths/{id}": {
			"get": {
				"operationId": "GetRegistrationPath",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseEnvelope_RegistrationPathDetail_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": 200,
											"error": false,
											"message": "Success",
											"data": {
												"id": 1,
												"general": {
													"name": "Path 1",
													"level": "0",
													"registration_type": "1",
													"description": "This is the first path",
													"start_date": "2021-01-01T00:00:00Z",
													"end_date": "2021-01-01T00:00:00Z"
												},
												"configuration": [
													{
														"id": 1,
														"fields": [
															{
																"field_type": "2",
																"is_required": true,
																"options": [
																	{
																		"id": 1,
																		"label": "Option 1"
																	},
																	{
																		"id": 2,
																		"label": "Option 2"
																	}
																],
																"ordering": 1,
																"title": "Field 1",
																"id": 1,
																"is_editable": true,
																"point": null
															}
														],
														"name": "General",
														"is_deleteable": true,
														"is_editable": true
													}
												],
												"schools": [
													1,
													2,
													3,
													4
												],
												"payment_config": {
													"items": [
														{
															"billing_schedule_type": "1",
															"name": "Item 1",
															"price": 1000000,
															"id": 1
														}
													],
													"payment_methods": [
														"BANK_TRANSFER"
													]
												},
												"zone_id": null,
												"has_applicant": false
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get a Registration Path",
				"summary": "Get a registration path",
				"tags": [
					"Registration Path"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRegistrationPath",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseEnvelope_RegistrationPath_"
								}
							}
						}
					}
				},
				"description": "Update a Registration Path",
				"summary": "Update a registration path",
				"tags": [
					"Registration Path"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegistrationPathPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRegistrationPath",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseEnvelope_null_"
								}
							}
						}
					},
					"400": {
						"description": "Registration Path already have applicants",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseEnvelope_null_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": 400,
											"error": true,
											"message": "Registration Path already have applicants",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete a Registration Path",
				"summary": "Delete a registration path",
				"tags": [
					"Registration Path"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v2/ppdb/registration-paths/{registrationPathId}/registration-payments": {
			"get": {
				"operationId": "GetRegistrationPayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseEnvelope_RegistrationPathPayment_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": 200,
											"error": false,
											"message": "Success",
											"count": 100,
											"next": "",
											"previous": "",
											"data": [
												{
													"id": 1,
													"billing_schedule_type": "1",
													"student_name": "John Doe",
													"fee_names": [
														"Item 1"
													],
													"payment_amount": 1000000,
													"payment_date": "2021-01-01T00:00:00Z",
													"payment_method": "Bank Transfer",
													"payment_status": "COMPLETE",
													"transaction_code": "123456"
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get All Registration Payments paginated",
				"summary": "Get all registration payments",
				"tags": [
					"Registration Path"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "registrationPathId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page_size",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "billing_schedule_type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RegistrationPathBillingScheduleType"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}